+++ bluetooth_low_energy/lib/bluetooth_low_energy.dart
@@ -1,5 +1,11 @@
+/// A Flutter plugin for controlling the bluetooth low energy, supports central
+/// and peripheral apis.
+library;
+
 export 'package:bluetooth_low_energy_platform_interface/bluetooth_low_energy_platform_interface.dart'
     hide
+        MyCentralManager,
+        MyPeripheralManager,
         MyObject,
         MyCentral,
         MyPeripheral,