+++ bluetooth_low_energy_platform_interface/lib/src/my_central_manager.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/cupertino.dart';
+import 'package:log_service/log_service.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+
+import 'central_manager.dart';
+
+/// Platform-specific implementations should implement this class to support
+/// [CentralManager].
+abstract class MyCentralManager extends PlatformInterface
+    with LoggerController
+    implements CentralManager {
+  static final Object _token = Object();
+
+  static MyCentralManager? _instance;
+
+  /// The default instance of [MyCentralManager] to use.
+  static MyCentralManager get instance {
+    final instance = _instance;
+    if (instance == null) {
+      throw UnimplementedError(
+        'CentralManager is not implemented on this platform.',
+      );
+    }
+    return instance;
+  }
+
+  /// Platform-specific implementations should set this with their own
+  /// platform-specific class that extends [MyCentralManager] when
+  /// they register themselves.
+  static set instance(MyCentralManager instance) {
+    PlatformInterface.verifyToken(instance, _token);
+    _instance = instance;
+  }
+
+  /// Constructs a [MyCentralManager].
+  MyCentralManager() : super(token: _token);
+
+  @protected
+  @override
+  Logger get logger => Logger('$CentralManager');
+}